Create README.md
The aim of this project is to perform an in-depth exploration, cleaning, and analysis of the Coffee Quality Database to extract valuable insights related to coffee production and quality. The primary objectives are to clean the data by handling missing values and inconsistencies, then conduct exploratory data analysis (EDA) through univariate, bivariate, and multivariate techniques. Visualizations will be created to illustrate the distribution, relationships, and patterns in the dataset.